requirements.txt
affine==2.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.12
botocore==1.38.12
branca==0.8.1
cachelib==0.13.0
cachetools==5.5.2
cartopy==0.24.1
certifi==2025.4.26
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
contextily==1.6.2
contourpy==1.3.2
cramjam==2.9.1
cryptography==44.0.3
cycler==0.12.1
dash==3.0.4
dash-extensions==2.0.4
dash-leaflet==1.1.3
dash-svg==0.0.12
dask==2025.4.1
dataclass-wizard==0.35.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
eccodes==2.41.0
editorconfig==0.17.0
elevation==1.1.3
elevatr==0.3.0
eodag==3.4.0
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.17.0
findlibs==0.1.1
flask==3.0.3
flask-caching==2.3.1
folium==0.19.5
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
geobuf==2.0.0
geographiclib==2.0
geojson==3.2.0
geomet==1.1.0
geopandas==1.0.1
geopy==2.4.1
h11==0.16.0
h5py==3.13.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==9.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsbeautifier==1.15.4
json5==0.12.0
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lazy-loader==0.4
lightning-utilities==0.13.1
locket==1.0.0
lxml==5.4.0
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mercantile==1.2.1
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netcdf4==1.7.2
networkx==3.4.2
notebook-shim==0.2.4
numpy==2.2.3
numpy-stl==3.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
odc-geo==0.4.10
odc-loader==0.5.1
odc-stac==0.4.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
ply==3.11
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.0
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyjwt==2.10.1
pykrige==1.7.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
pystac==1.13.0
pystac-client==0.8.6
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-utils==3.9.1
pytorch-lightning==2.5.0.post0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rioxarray==0.19.0
rpds-py==0.24.0
s3transfer==0.12.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
setuptools==75.8.2
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
stream-zip==0.0.83
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.5.10
tinycss2==1.4.0
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.2
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
whoosh==2.7.4
widgetsnbextension==4.0.14
xarray==2025.4.0
xyzservices==2025.4.0
yarl==1.18.3
zipp==3.21.0
